Animation update.
Renamed some poses.
One-handed pose now has smoother transitions for sprinting.
Off-hand fix for pistol poses.
More sprinting restrictions.
Grenades are now thrown with half velocity if crouching.

+++ src/main/java/com/mrcrayfish/guns/client/render/pose/WeaponPose.java
@@ -5,8 +5,6 @@ import com.mojang.math.Vector3f;
 import com.mrcrayfish.guns.client.handler.GunRenderingHandler;
 import com.mrcrayfish.guns.client.handler.ReloadHandler;
 import com.mrcrayfish.guns.client.render.IHeldAnimation;
-import com.mrcrayfish.guns.init.ModSyncedDataKeys;
-
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.renderer.MultiBufferSource;
@@ -54,7 +52,7 @@ public abstract class WeaponPose implements IHeldAnimation
     protected abstract AimPose getDownPose();
 
     /**
-     * Gets the pose of the player when looking directly down
+     * If this weapon is raised when sprinting
      */
     public boolean doRaiseWhenSprint() {
 		return false;
@@ -77,8 +75,8 @@ public abstract class WeaponPose implements IHeldAnimation
         ModelPart mainArm = right ? rightArm : leftArm;
         ModelPart secondaryArm = right ? leftArm : rightArm;
         
-        float reloadProgress = (float) ReloadHandler.get().getReloadProgress(Minecraft.getInstance().getFrameTime());
-        float sprintTransition = (float) GunRenderingHandler.get().getSprintTransition(Minecraft.getInstance().getFrameTime());
+        float reloadProgress = ReloadHandler.get().getReloadProgress(Minecraft.getInstance().getFrameTime());
+        float sprintTransition = GunRenderingHandler.get().getSprintTransition(Minecraft.getInstance().getFrameTime());
         float angle = Mth.lerp(sprintTransition, this.getPlayerPitch(player), (doRaiseWhenSprint() ? -0.4F : 0.3F));
         angle = Mth.lerp(reloadProgress, angle, 0.1F);
         float angleAbs = Math.abs(angle);
